Mac compatability initial commit (notes in commit message)
Several files modified to allow for os-specific commands and file/directory references. In python, this was done through consistent use of the os.path commands, in GAMS it was through consistent use of %ds% global, and in R it was through file.path function calls. Several file/call flow modifications are included here with more details in the list below.

The tests completed for this branch used:
macOS Mojave v10.14.6
Anaconda with Python 3.7.4
R 3.6.1 (installed alongside Anaconda)
GAMS 24.7

The main difference in running the model, relative to windows, is that the cases do not pop up in new command prompts and instead run in the background. Said differently, on Mac machines the model instances run quietly with minimal status updates printed to the screen and the user should investigate the cases' gamslog.txt files to observe progress. The reason for this is there is no succinct means to launch a new terminal on a Mac as compared to the 'start /wait ...' command available in Windows. Therefore, I get around this by creating a python subprocess.Popen instance that runs in the background.

A few caveats (several to be added to the ReadMe and User Guide):
1. The user should either a) use the GAMS terminal application that is installed alongside GAMS -or- permanently add their GAMS system directory to the PATH environment variable (similar to windows, e.g. export PATH="gams/sysdir":$PATH )
2. The solver is now specified in cases.csv - allowing the user to quickly choose between cplex/gurobi/...; this was previously os-specific given our usage of gurobi on the HPC
3. The user might need to specify the R_LIBS_USER environment variable similar to the adjustment of the path environment variable (e.g. export R_LIBS_USER=/my/R/library/directory/) - this is typically only true if the user does not have admin rights
4. The installation of the gdxrrw package in packagesetup.r is now os-specific and uses gdxrrw_1.0.5.tgz as the installation file on mac/linux machines (the installation file is being pushed with this commit)
5. reflow bash/command prompt output is now written to gamslog.txt (see changes to d_callreflow) - note the logoption is set to '2' for Mac/Linux so that it does not print to the screen during the call to runbatch.py, keeping the screen clean of anything besides what is purposefully printed by python
6. Small change to LDC_prep.py that did not cause issues in Windows but caused several issues on a Mac - spoke to WF who did not see an issue with this change
7. No longer creating the outputs/variabilityFiles folder in d_solveprep.gms as this was already done in runbatch.py
8. reflowbatch.py calls to the bash mimic the adjustments made in runbatch.py (subprocess.Popen...)
9. Adding notes to runbatch.py to indicate that for the intertemporal and window cases you'll need to specify GAMSDir in the setupEnvironment() function of runbatch.py
10. Removing calls to bokeh at the end of e_report_dump and having them written to the call_...sh/bat file via runbatch.py
11. (unrelated patch) - Scaling emit_tax by emit_scale
12. (cautionary note) - We will soon need to update/adjust our usage of the reshape2 package in R -- some of the commands are to be deprecated soon and will begin to cause errors and not just warnings