Committing first version of USREP-ReEDS linkage.
This version works for both the benchmark and counterfactual cases. There are still some modifications necessary (see '!!!!' throughout code) but these are mostly just for operational improvements (e.g. automating USREP reporting) and tidiness. An example of a benchmark and counterfactual case is presented in cases_ur.csv with 'Ref' being the benchmark case and 'cf_Ref' being the reproduction of that benchmark case. The linkage_gdx folder contains the necessary data to replicate that reference case and cf_ref.